requirements.txt
flask
flask_sqlalchemy
flask_login
werkzeug
psycopg2-binary